QualityAssurance/requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.2.1
django-filter==2.1.0
django-redis-sessions==0.6.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
httpie==1.0.2
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.2.post1
Pygments==2.4.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
